Assignment-1
we start with a backgroundColor state variable set to 'white'. The toggleBackgroundColor function is used to toggle the background color. When the button is clicked, it checks the current color and updates the state to either 'lightblue' or 'white' using a conditional (ternary) operator.

The background color is applied to the <div> element using the inline style prop. The height, display, alignItems, and justifyContent styles are added to center the button vertically and horizontally.

To use this component, create a new React application and replace the contents of the default src/App.js file with the code above. Save the file, and the application should automatically update with the new background color when the button is clicked.